requirements.txt
transformers
torch
langchain
psycopg2-binary
sentence-transformers
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyPDF2
python-docx